T746:orkaudio: replace ACE cross-platform compatibility layer
The general idea is that ACE is not as popular as APR and is showing
its age. Linux distributions do not package ACE by default, so it has
to be built every time orkaudio is deployed to a new Linux distribution.

ACE functionality is being replaced by native C++11 features, and
by the Apache Runtime library(APR).

APR powers various open source projects such as apache, freeswitch,
log4cxx and subversion.

Features that are being replaced by native C++11:
    semaphores
    mutexes
    threading
    file operations
Features that are being replaced by APR:
    timers
    networking
    dynamic shared object loading (loading .dll and .so files)