Improving deterministic mapping swap decision process
Since the swap decision process was getting more complicated we
decided to move parts of it into ShouldSwapChannels function.
Previously we were trying to match the ipstream with the localip
address. This was actually not enough since we found out that the rtp
stream could be different than the localip address.In order to
cover these cases we also started checking against the
media gateways. If the source is from a media gateways it is
not local. The original version of this patch is in T751_B