T786:orkaudio: Add HTTPS client and server
HTTPS server will accept and process orkaudio
commands except via a TLS session. This protects
against passive eavesdropping: the remote client
making the request is not verified (yet).

HTTPS client supports connection to a remote
orktrack via TLS. Again, the goal is to address
eavesdropping -- the orkaudio cleint does not
send client-side certificates to validate
to the remote (yet). Persistent connections
are supported via HTTP Keep-Alive. To use
HTTPS to connect to a remote tracker, the
TrackerHostname should start with https://

This support is NOT added to Centos6 and Windows.